API/requirements.txt
appdirs==1.4.4
astroid==2.6.6
autopep8==1.5.7
black==21.6b0
blinker==1.4
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-request-params==0.3.0
flask-requests==0.0.14
greenlet==1.1.0
idna==3.2
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pampy==0.3.0
pathspec==0.8.1
pdfkit==0.6.1
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.9.6
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
regex==2021.4.4
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.18
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
